Implement token auth for Envoy firmware v7
When a request is made, the token's freshness is checked - if it's less than two minutes away from expiring, the token is refreshed.

The DIY installer token is fetched by emulating the Entrez UI, based on great work done by @brodykenrick in https://github.com/brodykenrick/amber-enphase-zero-export-switcher-tool/tree/for-envoy-firmware-v7

Readme updated to include instructions for v7.

All the code is now compatible with both firmware v5 and v7, including the telemetry fetching